Covert PermissionManager to api
