+++ platform/android/api/api.h
@@ -1,0 +1,2 @@
+void register_android_api();
+void unregister_android_api();
